fix circular dependence
